Move Default Vehicle Analysis to separate Files
