+++ RetireSimple.Engine/Data/InvestmentVehicle/InvestmentVehicleBase.cs
@@ -38,9 +38,10 @@ namespace RetireSimple.Engine.Data.InvestmentVehicle {
 		[JsonIgnore, NotMapped]
 		public static readonly OptionsDict DefaultInvestmentVehicleOptions =
 			new() {
-				["VehicleTaxPercentage"] = "0.29", //Tax to apply on contribution/withdrawal/valuation
-				["VehicleMonthlyContribution"] = "0.00m",
-				["VehicleContributionInvestmentTarget"] = "-1" ///Default, simulates the <see cref="CashHoldings"/> contributions
+				["analysisLength"] = "60",
+				["vehicleTaxPercentage"] = "0.29", //Tax to apply on contribution/withdrawal/valuation
+				["cashContribution"] = "0.00m",
+				["vehicleContributionInvestmentTarget"] = "-1" ///Default, simulates the <see cref="CashHoldings"/> contributions
 			};
 
 		/// <summary>
@@ -60,8 +61,8 @@ namespace RetireSimple.Engine.Data.InvestmentVehicle {
 			//TODO Pre-process expenses/transfers (from investment in vehicle) to adjust for taxes
 			// This will be more relevant when expenses become a factor in analysis.
 
-			var preTaxModel = GeneratePreTaxModels(options, containedModels, cashSim);
-			var postTaxModel = GeneratePostTaxModels(options, containedModels, cashSim);
+			var preTaxModel = GeneratePreTaxModels(combinedOptions, containedModels, cashSim);
+			var postTaxModel = GeneratePostTaxModels(combinedOptions, containedModels, cashSim);
 
 			var newModel =
 				new InvestmentVehicleModel(InvestmentVehicleId, preTaxModel, postTaxModel);
@@ -80,138 +81,25 @@ namespace RetireSimple.Engine.Data.InvestmentVehicle {
 			List<InvestmentModel> models,
 			List<decimal>? cashContribution = null);
 
+
+		//Internal Methods for use in the template method
 		private List<InvestmentModel> GetContainedInvestmentModels(OptionsDict options) {
 			var models = new List<InvestmentModel>();
 			foreach (var investment in Investments) {
 				models.Add(investment.InvokeAnalysis(options));
 			}
-
 			return models;
 		}
 
 		private OptionsDict MergeOverrideOptions(OptionsDict manualOptions) {
 			var mergedOptions = new OptionsDict(manualOptions);
-
 			foreach (var key in AnalysisOptionsOverrides.Keys) {
 				mergedOptions.TryAdd(key, AnalysisOptionsOverrides[key]);
 			}
-
 			foreach (var key in DefaultInvestmentVehicleOptions.Keys) {
 				mergedOptions.TryAdd(key, DefaultInvestmentVehicleOptions[key]);
 			}
-
 			return mergedOptions;
-		}
-
-
-		///The following methods provide common logic for implementing the template method. You can
-		///have specific modules wrap around these if the logic applies (even partially).
-		public static InvestmentModel GeneratePreTaxModelDefaultAfterTaxVehicle(OptionsDict options,
-			List<InvestmentModel> models,
-			List<decimal>? cashContribution = null) {
-			//The logic here is a bit confusing at first, but here is an explanation of the transforms
-			var minModel = Enumerable.Range(0, models[0].MinModelData.Count) //Project the total length of the model with indexes
-									.Select(model =>                                    //For each index,
-										models.Select(m => m.MinModelData[model]).Sum()); //project each model's value at that index and sum
-			var maxModel = Enumerable.Range(0, models[0].MaxModelData.Count)
-									.Select(model =>
-										models.Select(m => m.MaxModelData[model]).Sum());
-			var avgModel = Enumerable.Range(0, models[0].AvgModelData.Count)
-									.Select(model =>
-										models.Select(m => m.AvgModelData[model]).Sum());
-
-			//If cash-based contributions exist, transform all models to include them
-			if (cashContribution != null) {
-				minModel = minModel.Select((val, idx) => val + cashContribution[idx]);
-				maxModel = maxModel.Select((val, idx) => val + cashContribution[idx]);
-				avgModel = avgModel.Select((val, idx) => val + cashContribution[idx]);
-			}
-
-			var preTaxModel = new InvestmentModel() {
-				MinModelData = minModel.ToList(),
-				MaxModelData = maxModel.ToList(),
-				AvgModelData = avgModel.ToList()
-			};
-
-			return preTaxModel;
-		}
-
-		public static InvestmentModel GeneratePostTaxModelDefaultAfterTaxVehicle(OptionsDict options,
-			List<InvestmentModel> models,
-			List<decimal>? cashContribution = null) {
-			//We are basically using the pretax model and applying tax to understand value after tax.
-			var preTaxModel =
-				GeneratePreTaxModelDefaultAfterTaxVehicle(options, models, cashContribution);
-
-			var taxPercentage = decimal.Parse(options["VehicleTaxPercentage"]);
-			var postTaxModel = new InvestmentModel() {
-				MinModelData = preTaxModel.MinModelData.Select(v => v * (1 - taxPercentage))
-					.ToList(),
-				MaxModelData = preTaxModel.MaxModelData.Select(v => v * (1 - taxPercentage))
-					.ToList(),
-				AvgModelData = preTaxModel.AvgModelData.Select(v => v * (1 - taxPercentage))
-					.ToList()
-			};
-			return postTaxModel;
-		}
-
-		public static InvestmentModel GeneratePreTaxModelDefaultPreTaxVehicle(OptionsDict options,
-			List<InvestmentModel> models,
-			List<decimal>? cashContribution = null) {
-
-			var taxPercentage = decimal.Parse(options["VehicleTaxPercentage"]);
-			//The logic here is a bit confusing at first, but here is an explanation of the transforms
-			var minModel = Enumerable.Range(0, models[0].MinModelData.Count)                //Project the total length of the model with indexes
-									.Select(model =>                                        //For each index,
-										models.Select(m => m.MinModelData[model]).Sum())    //project each model's value at that index and sum
-									.Select(v => v * (1 - taxPercentage));                  // Apply Taxes
-			var maxModel = Enumerable.Range(0, models[0].MaxModelData.Count)
-									.Select(model =>
-										models.Select(m => m.MaxModelData[model]).Sum())
-									.Select(v => v * (1 - taxPercentage));
-			var avgModel = Enumerable.Range(0, models[0].AvgModelData.Count)
-									.Select(model =>
-										models.Select(m => m.AvgModelData[model]).Sum())
-									.Select(v => v * (1 - taxPercentage));
-
-			//If cash-based contributions exist, transform all models to include them
-			//This version assumes that cash contributions are already taxed
-			if (cashContribution != null) {
-				minModel = minModel.Select((val, idx) => val + cashContribution[idx]);
-				maxModel = maxModel.Select((val, idx) => val + cashContribution[idx]);
-				avgModel = avgModel.Select((val, idx) => val + cashContribution[idx]);
-			}
-
-			var preTaxModel = new InvestmentModel() {
-				MinModelData = minModel.ToList(),
-				MaxModelData = maxModel.ToList(),
-				AvgModelData = avgModel.ToList()
-			};
-
-			return preTaxModel;
-		}
-
-		public List<decimal> SimulateCashContributionsDefaultAfterTax(OptionsDict options) {
-			var analysisLength = int.Parse(options["AnalysisLength"]);
-			var cashContribution = decimal.Parse(options["CashContribution"]);
-			var currentHoldings = CashHoldings;
-
-			return Enumerable.Range(0, analysisLength)
-				.Select((val, idx) => currentHoldings + cashContribution * idx)
-				.ToList();
-		}
-
-		public List<decimal> SimulateCashContributionsDefaultPreTax(OptionsDict options) {
-			var analysisLength = int.Parse(options["AnalysisLength"]);
-			var cashContribution = decimal.Parse(options["CashContribution"]);
-			var taxPercentage = decimal.Parse(options["VehicleTaxPercentage"]);
-			var currentHoldings = CashHoldings;
-
-			return Enumerable.Range(0, analysisLength)
-				.Select((val, idx) => cashContribution * idx)
-				.Select(val => val * (1 - taxPercentage))
-				.Select(val => val + currentHoldings)
-				.ToList();
 		}
 
 	}
