+++ RetireSimple.Engine/Data/InvestmentVehicle/VehicleRothIRA.cs
@@ -1,14 +1,16 @@
-﻿namespace RetireSimple.Engine.Data.InvestmentVehicle {
+﻿using RetireSimple.Engine.Analysis;
+
+namespace RetireSimple.Engine.Data.InvestmentVehicle {
 	public class VehicleRothIRA : InvestmentVehicleBase {
 		public override InvestmentModel GeneratePostTaxModels(OptionsDict options,
 													List<InvestmentModel> models,
 													List<decimal>? cashContribution = null)
-			=> GeneratePreTaxModelDefaultPreTaxVehicle(options, models, cashContribution);
+			=> VehicleDefaultAS.GeneratePreTaxModelDefaultPreTaxVehicle(options, models, cashContribution);
 		public override InvestmentModel GeneratePreTaxModels(OptionsDict options,
 															List<InvestmentModel> models,
 															List<decimal>? cashContribution = null)
-			=> GeneratePreTaxModelDefaultPreTaxVehicle(options, models, cashContribution);
+			=> VehicleDefaultAS.GeneratePreTaxModelDefaultPreTaxVehicle(options, models, cashContribution);
 		public override List<decimal> SimulateCashContributions(OptionsDict options)
-			=> SimulateCashContributionsDefaultPreTax(options);
+			=> VehicleDefaultAS.SimulateCashContributionsDefaultPreTax(this, options);
 	}
 }
